fix terminal not stopping on child exit
Ctrl-d or exit in the terminal did not close the window.

	TypeError: __tab_child_exited_cb() takes exactly 2 arguments (3 given)

First seen Ubuntu 15.10, where Vte 2.91 gives a new status parameter to
the child-exited signal.

Older Vte does not give a status parameter; Vte 2.90 on Ubuntu 14.04,
Vte 0.34.9 on Fedora 20, nor Vte 0.34.2 on Fedora 18.

Accept an optional status parameter, but ignore it.