Use the user's shell, rather than always BASH
Bash is a pretty average shell.  Some kids (or older people alike)
might like a pretty looking shell like fish.  We can read the
user's current shell from the SHELL envar, which is set through the
"chsh" command.  Eventually, we could provide a GUI for setting that,
but that is for the future.