Font size, keep size on theme change
When theme was changed between dark or light, the font size would reset
back to default.

During VT reconfigure, keep the font size as it was, by allowing the new
font to default to the previous font size.

Also make a font size default instead of inheriting the Pango default.